+++ src/components/NautobotGrid/index.ts
@@ -1,0 +1,2 @@
+export { default } from './NautobotGrid';
+export type { NautobotGridProps } from './NautobotGrid';
