--- src/components/NautobotGridLayout/index.ts
@@ -1,2 +1,0 @@
-export { default } from './NautobotGridLayout';
-export type { NautobotGridLayoutProps } from './NautobotGridLayout';
